Refactor previews
