+++ app/src/main/java/com/ashipo/metropolitanmuseum/ui/util/SharedScopes.kt
@@ -1,12 +1,23 @@
-package com.ashipo.metropolitanmuseum.ui
+@file:OptIn(ExperimentalSharedTransitionApi::class)
+
+package com.ashipo.metropolitanmuseum.ui.util
 
 import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.AnimatedVisibilityScope
 import androidx.compose.animation.ExperimentalSharedTransitionApi
 import androidx.compose.animation.SharedTransitionLayout
+import androidx.compose.animation.SharedTransitionScope
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
+import com.ashipo.metropolitanmuseum.ui.LocalAnimatedVisibilityScope
+import com.ashipo.metropolitanmuseum.ui.LocalSharedTransitionScope
 
-@OptIn(ExperimentalSharedTransitionApi::class)
+/**
+ * A wrapper composable that provides [SharedTransitionScope] and [AnimatedVisibilityScope] for the
+ * [content]. Can be used in previews or tests.
+ *
+ * @param content that needs the scopes
+ */
 @Composable
 fun SharedScopes(content: @Composable () -> Unit) {
     SharedTransitionLayout {
